+++ x/perp/keeper/grpc_query.go
@@ -23,9 +23,37 @@ func NewQuerier(k Keeper) types.QueryServer {
 
 var _ types.QueryServer = queryServer{}
 
-func (q queryServer) QueryTraderPosition(
-	goCtx context.Context, req *types.QueryTraderPositionRequest,
-) (*types.QueryTraderPositionResponse, error) {
+func (q queryServer) QueryPositions(
+	goCtx context.Context, req *types.QueryPositionsRequest,
+) (*types.QueryPositionsResponse, error) {
+	if req == nil {
+		return nil, status.Error(codes.InvalidArgument, "invalid request")
+	}
+	_, err := sdk.AccAddressFromBech32(req.Trader) // just for validation purposes
+	if err != nil {
+		return nil, err
+	}
+
+	ctx := sdk.UnwrapSDKContext(goCtx)
+
+	pools := q.k.VpoolKeeper.GetAllPools(ctx)
+	var positions []*types.QueryPositionResponse
+
+	for _, pool := range pools {
+		position, err := q.position(ctx, pool.Pair, req.Trader)
+		if err == nil {
+			positions = append(positions, position)
+		}
+	}
+
+	return &types.QueryPositionsResponse{
+		Positions: positions,
+	}, nil
+}
+
+func (q queryServer) QueryPosition(
+	goCtx context.Context, req *types.QueryPositionRequest,
+) (*types.QueryPositionResponse, error) {
 	if req == nil {
 		return nil, status.Error(codes.InvalidArgument, "invalid request")
 	}
@@ -41,10 +69,10 @@ func (q queryServer) QueryTraderPosition(
 		return nil, err
 	}
 
-	return q.traderPosition(ctx, pair, req.Trader)
+	return q.position(ctx, pair, req.Trader)
 }
 
-func (q queryServer) traderPosition(ctx sdk.Context, pair common.AssetPair, trader string) (*types.QueryTraderPositionResponse, error) {
+func (q queryServer) position(ctx sdk.Context, pair common.AssetPair, trader string) (*types.QueryPositionResponse, error) {
 	position, err := q.k.Positions.Get(ctx, keys.Join(pair, keys.String(trader)))
 	if err != nil {
 		return nil, err
@@ -67,7 +95,7 @@ func (q queryServer) traderPosition(ctx sdk.Context, pair common.AssetPair, trad
 		marginRatioIndex = sdk.Dec{}
 	}
 
-	return &types.QueryTraderPositionResponse{
+	return &types.QueryPositionResponse{
 		Position:         &position,
 		PositionNotional: positionNotional,
 		UnrealizedPnl:    unrealizedPnl,